klout score and whatnot

+++ db/migrate/20120623234014_create_repos.rb
@@ -2,7 +2,7 @@ class CreateRepos < ActiveRecord::Migration
   def change
     create_table :repos do |t|
       t.string :name
-      t.integer :vote_sum
+      t.float :vote_sum
       t.date :week_start
 
       t.timestamps
